+++ src/Algorithms/NoPID.cpp
@@ -1,7 +1,7 @@
 #include "Algorithms/NoPID.h"
 
 // Doesn't matter what the error is. Always run at max speed, stored in KP
-float NoPID::tick(float error) {
+double NoPID::tick(double error) {
 
     handleEndCondition(error);
